Add image for CPU usage

+++ project_setup.py
@@ -55,7 +55,7 @@ def setup(
     # Build the image:
     if build_image:
         print("Building default image for the demo:")
-        _build_image(project=project)
+        _build_image(project=project, with_gpu=gpus)
 
     # Set the secrets:
     _set_secrets(
@@ -83,27 +83,48 @@ def setup(
     return project
 
 
-def _build_image(project: mlrun.projects.MlrunProject):
-    assert project.build_image(
-        base_image="mlrun/mlrun-gpu",
-        commands=[
-            # Update apt-get to install ffmpeg (support audio file formats):
-            "apt-get update -y && apt-get install ffmpeg -y",
-            # Install demo requirements:
-            "pip install transformers==4.44.1",
-            "pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118",
-            "pip install bitsandbytes==0.41.1 accelerate==0.24.1 datasets==2.14.6 peft==0.5.0 optimum==1.13.2",
-            "pip install auto-gptq==0.4.2 --extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/",
-            "pip install langchain==0.0.327 openai==0.28.1",
-            "pip install git+https://github.com/suno-ai/bark.git",  # suno-bark
-            "pip install streamlit==1.28.0 st-annotated-text==4.0.1 spacy==3.7.2 librosa==0.10.1 presidio-anonymizer==2.2.34 presidio-analyzer==2.2.34 nltk==3.8.1 flair==0.13.0",
-            "python -m spacy download en_core_web_lg",
-            "pip install -U SQLAlchemy",
-            "pip uninstall -y onnxruntime-gpu onnxruntime",
-            "pip install onnxruntime-gpu",
-        ],
-        set_as_default=True,
-    )
+def _build_image(project: mlrun.projects.MlrunProject, with_gpu):
+    if with_gpu:
+        assert project.build_image(
+            base_image="mlrun/mlrun-gpu",
+            commands=[
+                # Update apt-get to install ffmpeg (support audio file formats):
+                "apt-get update -y && apt-get install ffmpeg -y",
+                # Install demo requirements:
+                "pip install transformers==4.44.1",
+                "pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cu118",
+                "pip install bitsandbytes==0.41.1 accelerate==0.24.1 datasets==2.14.6 peft==0.5.0 optimum==1.13.2",
+                "pip install auto-gptq==0.4.2 --extra-index-url https://huggingface.github.io/autogptq-index/whl/cu118/",
+                "pip install langchain==0.0.327 openai==0.28.1",
+                "pip install git+https://github.com/suno-ai/bark.git",  # suno-bark
+                "pip install streamlit==1.28.0 st-annotated-text==4.0.1 spacy==3.7.2 librosa==0.10.1 presidio-anonymizer==2.2.34 presidio-analyzer==2.2.34 nltk==3.8.1 flair==0.13.0",
+                "python -m spacy download en_core_web_lg",
+                "pip install -U SQLAlchemy",
+                "pip uninstall -y onnxruntime-gpu onnxruntime",
+                "pip install onnxruntime-gpu",
+            ],
+            set_as_default=True,
+        )
+    else:
+        assert project.build_image(
+            base_image="mlrun/mlrun",
+            commands=[
+                # Update apt-get to install ffmpeg (support audio file formats):
+                "apt-get update -y && apt-get install ffmpeg -y",
+                # Install demo requirements:
+                "pip install transformers==4.44.1",
+                "pip install torch==2.1.2 torchvision==0.16.2 torchaudio==2.1.2 --index-url https://download.pytorch.org/whl/cpu",
+                "pip install bitsandbytes==0.41.1 accelerate==0.24.1 datasets==2.14.6 peft==0.5.0 optimum==1.13.2",
+                "pip install auto-gptq==0.4.2",
+                "pip install langchain==0.0.327 openai==0.28.1",
+                "pip install git+https://github.com/suno-ai/bark.git",  # suno-bark
+                "pip install streamlit==1.28.0 st-annotated-text==4.0.1 spacy==3.7.2 librosa==0.10.1 presidio-anonymizer==2.2.34 presidio-analyzer==2.2.34 nltk==3.8.1 flair==0.13.0",
+                "python -m spacy download en_core_web_lg",
+                "pip install -U SQLAlchemy",
+                "pip install onnxruntime",
+            ],
+            set_as_default=True,
+        )
 
 
 def _set_secrets(
